% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wlts.R
\name{describe_collection}
\alias{describe_collection}
\title{...}
\usage{
describe_collection(URL, collection_id, ...)
}
\arguments{
\item{URL}{a \code{character} ...}

\item{collection_id}{a \code{character} ...}

\item{...}{a \code{character} ...}
}
\value{

}
\description{
...
}
