% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/request.R
\name{.build_url}
\alias{.build_url}
\title{}
\usage{
.build_url(URL, path, ...)
}
\arguments{
\item{URL}{\code{character} ...}

\item{path}{\code{character} ...}

\item{...}{\code{character} ...}

\item{query}{\code{character} ...}
}
\description{

}
