% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tibble.R
\name{.build_wlts_tb}
\alias{.build_wlts_tb}
\title{Build wlts tibble}
\usage{
.build_wlts_tb(result, index)
}
\arguments{
\item{result}{a named \code{list} from request in get_trajectory.}

\item{index}{a \code{numeric} point identifier.}
}
\value{
a \code{tibble} with trajectory data.
}
\description{
Build wlts tibble
}
