% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.check_datetime}
\alias{.check_datetime}
\title{...}
\usage{
.check_datetime(start_date, end_date)
}
\arguments{
\item{a}{\code{character} ...}

\item{...}{\code{character} ...}
}
\value{

}
\description{
...
}
