% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wlts.R
\name{get_trajectory}
\alias{get_trajectory}
\title{...}
\usage{
get_trajectory(
  URL,
  latitude,
  longitude,
  collections = NULL,
  start_date = NULL,
  end_date = NULL,
  ...
)
}
\arguments{
\item{latitude}{a \code{numeric} ...}

\item{longitude}{a \code{numeric} ...}

\item{collections}{a \code{character} ...}

\item{...}{a \code{character} ...}
}
\value{

}
\description{
...
}
