% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/request.R
\name{list_collections}
\alias{list_collections}
\title{...}
\usage{
list_collections(url = "http://brazildatacube.dpi.inpe.br/dev/wlts", ...)
}
\arguments{
\item{a}{\code{character} ...}
}
\value{

}
\description{
...
}
