% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mt_500_mapbiomas_cerrado}
\alias{mt_500_mapbiomas_cerrado}
\title{Trajectory data from Mato Grosso of collection Map Biomas 5}
\format{
A \code{wlts} tibble with 5.5K samples. A \code{wlts} tibble contains
the following meta-data:
query (a NULL value),
result (a tibble with trajectory information.).
}
\usage{
data("mt_500_mapbiomas_cerrado")
}
\description{
A dataset containing a wlts tibble, with result and query attributes.
}
\keyword{datasets}
