% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.wlts}
\alias{plot.wlts}
\alias{plot.alluvial}
\title{Plots for rwlts package}
\usage{
\method{plot}{wlts}(x, ..., type = "alluvial")

\method{plot}{alluvial}(x, ..., show_count = FALSE)
}
\arguments{
\item{x}{a \code{tibble} object from class \code{wlts}.}

\item{...}{additional functions}

\item{type}{a \code{character} with the type of plot.
Nowadays, only the "alluvial" plot is supported. By default is "alluvial".}

\item{show_count}{a \code{logical} parameter, if true, is added
the number of points on each bar. The default value is FALSE.}
}
\value{
a \code{gg} object from ggplot2 package.
}
\description{
A set of visualization methods for trajectories extracted in
 rwlts package.
}
\examples{
\dontrun{
 wlts_bdc <- "https://brazildatacube.dpi.inpe.br/wlts/"

 wlts_tibble <- get_trajectory(
                     URL        = wlts_bdc,
                     latitude    = c(-12, -11),
                     longitude   = c(-54, -55),
                     collections = "mapbiomas_amazonia-v5",
                     start_date  = "2015-07-01",
                     end_date    = "2017-07-01",
                     config = httr::add_headers("x-api-key" = "BDC-KEY"))

 plot(wlts_tibble)
}

}
