% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wltr-package.R
\docType{package}
\name{wlts-package}
\alias{wlts-package}
\alias{_PACKAGE}
\alias{wlts}
\title{wltr}
\description{
An R client to the web land trajectory service (WLTS)
}
\section{WLTS API}{


Implements an R interface to a web land trajectory service (WLTS)
that offers a land trajectory time series of remote sensing samples API.

The WLTS API has three commands:
\itemize{
   \item `list_collections`: lists all available collections on WLTS Service
   \item `trajectory`: Returns the Land Use and Land Cover for a spatio-temporal location
   \item `describe_collection`: retrieves the metadata of a collection
}
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/brazil-data-cube/rwlts/}
  \item Report bugs at \url{https://github.com/brazil-data-cube/rwlts/issues}
}

}
\author{
\strong{Maintainer}: Brazil Data Cube Team \email{brazildatacube@inpe.br}

Other contributors:
\itemize{
  \item National Institute for Space Research (INPE) [copyright holder]
}

}
