% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/request.R
\name{.build_url}
\alias{.build_url}
\title{Create a valid URL}
\usage{
.build_url(URL, path, query = NULL, names_list = NULL)
}
\arguments{
\item{URL}{a \code{character} URL of the WLTS Service.}

\item{path}{a \code{character} with URL path.}

\item{query}{a \code{list} A named list with HTTP query parameters.}

\item{names_list}{a \code{character} vector with the names of query
attributes.}
}
\value{
a \code{list} with \code{url} and \code{query} attributes.
}
\description{
Checks if the URL has patterns that may cause query problems.
}
