% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.check_datetime}
\alias{.check_datetime}
\title{Validates the dates}
\usage{
.check_datetime(start_date, end_date)
}
\arguments{
\item{start_date}{\code{character} Start date in RFC 3339 format}

\item{end_date}{\code{character} End date in RFC 3339 format}
}
\description{
Checks if the dates are in RFC 3339 format and if they are in a
 valid start and end date relationship
}
