% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.check_location}
\alias{.check_location}
\title{...}
\usage{
.check_location(latitude, longitude)
}
\arguments{
\item{a}{\code{character} ...}

\item{...}{\code{character} ...}
}
\value{

}
\description{
...
}
