% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wlts.R
\name{get_trajectory}
\alias{get_trajectory}
\title{Get Trajectory}
\usage{
get_trajectory(
  URL,
  latitude,
  longitude,
  collections = NULL,
  start_date = NULL,
  end_date = NULL,
  ...,
  query_info = FALSE
)
}
\arguments{
\item{URL}{a \code{character} URL of the WLTS Service.}

\item{latitude}{a \code{numeric} vector corresponding to latitude in
WGS84 coordinate system.}

\item{longitude}{a \code{numeric} vector corresponding to longitude in
WGS84 coordinate system.}

\item{collections}{a \code{character} vector of identifier (name) of
collections delimited by comma.}

\item{...}{a \code{list} Parameters to httr::GET function.}

\item{query_info}{a \code{bolean} flag, if true query information is
returned.}
}
\value{
a \code{object} of wlts class with query (if \code{query_info} is
 provided) and a tibble with trajectory requested.
}
\description{
Retrieves the land use and cover trajectories from the data
 collections given a location in space. The property \code{result} contains
 the feature identifier information, class, time, and the collection
 associated to the data item.
}
\examples{
\donttest{
 wlts_bdc <- "http://brazildatacube.dpi.inpe.br/dev/wlts/"

 get_trajectory(wlts_bdc, latitude = c(-12, -11.01), longitude = c(-54, -54),
               collections = "mapbiomas_amz_4_1")
}

}
