% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/request.R
\name{get_trajectory}
\alias{get_trajectory}
\title{...}
\usage{
get_trajectory(
  latitude,
  longitude,
  collections,
  url = "http://brazildatacube.dpi.inpe.br/dev/wlts",
  ...
)
}
\arguments{
\item{latitude}{a \code{character} ...}

\item{longitude}{a \code{character} ...}

\item{collections}{a \code{character} ...}

\item{...}{a \code{character} ...}
}
\value{

}
\description{
...
}
