% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wlts.R
\name{list_collections}
\alias{list_collections}
\title{List Collections}
\usage{
list_collections(URL, ...)
}
\arguments{
\item{URL}{a \code{character} URL of the WLTS Service.}

\item{...}{a \code{list} Parameters to httr::GET function}
}
\value{
a \code{character} vector with the available data collections.
}
\description{
Retrieves the list of available data collections.
}
\examples{
\donttest{
 wlts_bdc <- "https://brazildatacube.dpi.inpe.br/wlts/"

 list_collections(wlts_bdc)
}

}
