% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.wlts}
\alias{plot.wlts}
\title{Sankey Plot}
\usage{
\method{plot}{wlts}(wtls_tibble, type, ...)
}
\arguments{
\item{wtls_tibble}{a \code{tibble} object from class \code{wlts}.}

\item{show_count_transition}{a \code{logical} parameter, if true, is added
the number of points on each bar. The default value is FALSE.}
}
\value{
a \code{gg} object from ggplot2 package.
}
\description{
Visualization method based on the sankey graph. In which the
changes between each class are presented on the time axis.
}
\examples{
\donttest{
 wlts_bdc <- "https://brazildatacube.dpi.inpe.br/wlts/"

 wlts_tibble <- get_trajectory(
                     URL        = wlts_bdc,
                     latitude    = c(-12, -11),
                     longitude   = c(-54, -55),
                     collections = "mapbiomas5_amazonia",
                     start_date  = "2015-07-01",
                     end_date    = "2017-07-01",
                     config = httr::add_headers("x-api-key" = "BDC-KEY"))


 plot_sankey(wlts_tibble)
}

}
