% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/request.R
\name{request}
\alias{request}
\title{}
\usage{
request(URL, ..., query = NULL)
}
\arguments{
\item{URL}{\code{character} ...}

\item{...}{\code{character} ...}

\item{query}{\code{character} ...}
}
\value{

}
\description{

}
